requirements.txt
chardet
bs4
pytz
python-dotenv
pdfkit